+++ pyxel/calibration/calibration.py
@@ -23,10 +23,9 @@ from pyxel.calibration import (
     MyArchipelago,
 )
 from pyxel.calibration.fitting import ModelFitting
-from pyxel.observation.parameter_values import ParameterValues
-from pyxel.pipelines import ModelFunction, Processor
-
-from ..pipelines.processor import ResultType
+from pyxel.exposure import Readout
+from pyxel.observation import ParameterValues
+from pyxel.pipelines import FitnessFunction, Processor, ResultType
 
 try:
     import pygmo as pg
@@ -39,7 +38,6 @@ if t.TYPE_CHECKING:
     import pandas as pd
     import xarray as xr
 
-    from pyxel.exposure import Readout
     from pyxel.outputs import CalibrationOutputs
 
 
@@ -54,20 +52,16 @@ class Calibration:
     def __init__(
         self,
         outputs: "CalibrationOutputs",
-        readout: "Readout",
-        output_dir: t.Optional[Path] = None,
-        fitting: t.Optional[
-            ModelFitting
-        ] = None,  # Note: 'fitting' is set in '.run_calibration'
+        target_data_path: t.Sequence[Path],
+        fitness_function: FitnessFunction,
+        algorithm: Algorithm,
+        parameters: t.Sequence[ParameterValues],
+        readout: t.Optional["Readout"] = None,
         mode: Literal["pipeline", "single_model"] = "pipeline",
         result_type: Literal["image", "signal", "pixel"] = "image",
         result_fit_range: t.Optional[t.Sequence[int]] = None,
         result_input_arguments: t.Optional[t.Sequence[ParameterValues]] = None,
-        target_data_path: t.Optional[t.Sequence[Path]] = None,
         target_fit_range: t.Optional[t.Sequence[int]] = None,
-        fitness_function: t.Optional[ModelFunction] = None,
-        algorithm: t.Optional[Algorithm] = None,
-        parameters: t.Optional[t.Sequence[ParameterValues]] = None,
         pygmo_seed: t.Optional[int] = None,
         pipeline_seed: t.Optional[int] = None,
         num_islands: int = 1,
@@ -96,10 +90,7 @@ class Calibration:
         self._log = logging.getLogger(__name__)
 
         self.outputs = outputs
-        self.readout = readout
-
-        self._output_dir = output_dir  # type:t.Optional[Path]
-        self._fitting = fitting  # type: t.Optional[ModelFitting]
+        self.readout = readout if readout else Readout()  # type: Readout
 
         self._calibration_mode = CalibrationMode(mode)
 
@@ -120,8 +111,8 @@ class Calibration:
             target_fit_range if target_fit_range else []
         )  # type: t.Sequence[int]
 
-        self._fitness_function = fitness_function  # type: t.Optional[ModelFunction]
-        self._algorithm = algorithm  # type: t.Optional[Algorithm]
+        self._fitness_function = fitness_function  # type: FitnessFunction
+        self._algorithm = algorithm  # type: Algorithm
 
         self._parameters = (
             parameters if parameters else []
@@ -136,7 +127,7 @@ class Calibration:
         self._num_islands = num_islands  # type: int
         self._num_evolutions = num_evolutions  # type: int
         self._num_best_decisions = num_best_decisions  # type: t.Optional[int]
-        self._type_islands = Island(type_islands)  # type:Island
+        self._type_islands = Island(type_islands)  # type: Island
         self._pipeline_seed = pipeline_seed
         self._topology = (
             topology
@@ -149,32 +140,6 @@ class Calibration:
             weights_from_file
         )  # type: t.Optional[t.Sequence[Path]]
         self._weights = weights  # type: t.Optional[t.Sequence[float]]
-
-    @property
-    def output_dir(self) -> Path:
-        """TBW."""
-        if not self._output_dir:
-            raise RuntimeError("No 'output_dir' defined !")
-
-        return self._output_dir
-
-    @output_dir.setter
-    def output_dir(self, value: Path) -> None:
-        """TBW."""
-        self._output_dir = value
-
-    @property
-    def fitting(self) -> ModelFitting:
-        """TBW."""
-        if not self._fitting:
-            raise RuntimeError("No 'fitting' defined !")
-
-        return self._fitting
-
-    @fitting.setter
-    def fitting(self, value: ModelFitting) -> None:
-        """TBW."""
-        self._fitting = value
 
     @property
     def calibration_mode(self) -> CalibrationMode:
@@ -237,24 +202,18 @@ class Calibration:
         self._target_fit_range = value
 
     @property
-    def fitness_function(self) -> ModelFunction:
+    def fitness_function(self) -> FitnessFunction:
         """TBW."""
-        if not self._fitness_function:
-            raise RuntimeError("No 'fitness_function' defined !")
-
         return self._fitness_function
 
     @fitness_function.setter
-    def fitness_function(self, value: ModelFunction) -> None:
+    def fitness_function(self, value: FitnessFunction) -> None:
         """TBW."""
         self._fitness_function = value
 
     @property
     def algorithm(self) -> Algorithm:
         """TBW."""
-        if not self._algorithm:
-            raise RuntimeError("No 'algorithm' defined !")
-
         return self._algorithm
 
     @algorithm.setter
@@ -377,24 +336,25 @@ class Calibration:
         pg.set_global_rng_seed(seed=self.pygmo_seed)
         self._log.info("Pygmo seed: %d", self.pygmo_seed)
 
-        self.output_dir = output_dir
-
-        self.fitting = ModelFitting(
-            processor=processor, variables=self.parameters, readout=self.readout
-        )
-        self.fitting.configure(
-            calibration_mode=self.calibration_mode,
+        fitting = ModelFitting(
+            processor=processor,
+            variables=self.parameters,
+            readout=self.readout,
+            calibration_mode=CalibrationMode(self.calibration_mode),
+            simulation_output=ResultType(self.result_type),
             generations=self.algorithm.generations,
             population_size=self.algorithm.population_size,
-            simulation_output=self.result_type,
             fitness_func=self.fitness_function,
+            file_path=output_dir,
+        )
+
+        fitting.configure(
             target_output=self.target_data_path,
             target_fit_range=self.target_fit_range,
             out_fit_range=self.result_fit_range,
             input_arguments=self.result_input_arguments,
             weights=self.weights,
             weights_from_file=self.weights_from_file,
-            file_path=output_dir,
         )
 
         if self.num_islands > 1:  # default
@@ -417,7 +377,7 @@ class Calibration:
                 num_islands=self.num_islands,
                 udi=user_defined_island,
                 algorithm=self.algorithm,
-                problem=self.fitting,
+                problem=fitting,
                 pop_size=self.algorithm.population_size,
                 bfe=user_defined_bfe,
                 topology=topo,
@@ -433,7 +393,7 @@ class Calibration:
             )
 
             ds.attrs["topology"] = self.topology
-            ds.attrs["result_type"] = str(self.fitting.sim_output)
+            ds.attrs["result_type"] = str(fitting.sim_output)
 
         else:
             raise NotImplementedError("Not implemented for 1 island.")